+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 1.0.3:
+* Create a LocalePropertyProvider.
 * Import property providers and customized servlet compatible with MGWT framework.
 
 ## 1.0.2: